requirements.txt
paho-mqtt
grpcio
grpcio-tools
pymongo
flask